01platformCt/requirements.txt
APScheduler==3.10.4
backports.zoneinfo==0.2.1
Flask==2.2.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpath==0.82.2
Mako==1.2.4
MarkupSafe==2.1.3
paramiko==2.11.0
PyJWT==2.8.0
PyMySQL==1.1.0
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.7.1
Werkzeug==2.2.3
WTForms==3.0.1

